+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
     - Mandatory field for the rent increase capping limit
 - Rent increase
     - Generate notification document on rent increase
+- New document types
+    - cover letter
+    - signed contract
+    - signed rent increase
 
 ### Changed